moved text2hash to utils.ts
